requirements.txt
# --- Core API & Data Extraction (Copernicus Integration) ---
cdsapi>=0.7.7
ecmwf-datastores-client>=1.0.0

# --- Deep Learning & Physics Gating ---
torch>=2.0.0
scikit-learn>=1.3.0

# --- Data Processing & Resilient Tensors (Data Healing) ---
numpy>=1.24.0
pandas>=2.0.0
xarray>=2023.1.0
netCDF4>=1.6.0
h5netcdf>=1.2.0
scipy>=1.10.0

# --- Visualization & Geospatial Analysis ---
matplotlib>=3.7.0
seaborn>=0.12.0
cartopy>=0.22.0
shapely>=2.0.0
pyproj>=3.5.0

# --- Environment & Operational Utilities ---
python-dotenv>=1.0.0
tqdm>=4.65.0